Add files via upload
This project employs Naive Bayes as a baseline model and conducts a comparative analysis of RNN and DNN to identify the most effective models for detecting Gender-Based Violence (GBV) tweets.

The objective is to contribute to the eradication of GBV online, specifically technology-facilitated Gender-Based Violence. By accurately identifying GBV activity on social media, the system aims to facilitate reporting, protect victims, and assist law enforcement in the strategic deployment of resources.

Moreover, the application extends to the detection and prevention of intimate partner violence, enhancing the capability to prevent the recurrence of GBV offenses.